+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## neptune-xgboost 1.0.0
+
+### Changes
+- `NeptuneCallback` now accepts a namespace `Handler` as an alternative to `Run` for the `run` argument. This means that
+  you can call it like `NeptuneCallback(run=run["some/namespace/"])` to log everything to the `some/namespace/`
+  location of the run.
+
+### Breaking changes
+- Instead of the `log()` method, the integration now uses `append()` which is available since version 0.16.14
+  of neptune-client.
+
 ## neptune-xgboost 0.10.1
 
 ### Changes